+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [1.7.3-SNAPSHOT]
 
+- Fixed ViewModel validation bug for ComboBox, ChoiceBox and Spinner
+- Removed faulty choicebox builder and replaced it with one similar to the combobox builder
 - Autocomplete ComboBox listview matches width of ComboBox by default
 - `authInterceptor` was deprecated in favor of better named `requestInterceptor`
 - Android SDK compatibilty (See https://github.com/edvin/tornadofx-android-compat)